# openapi.yaml
openapi: 3.1.0
info:
  title: OCD Management API
  description: API for the OCD Management application - track symptoms, habits, exercises, and progress
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
paths:
...
  /api/symptoms:
    get:
      summary: List symptom logs
      tags: [Symptoms]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of symptom logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  symptoms:
                    type: array
                    items:
                      $ref: "#/components/schemas/Symptom"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Log symptoms
      tags: [Symptoms]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SymptomInput"
      responses:
        "201":
          description: Symptom log created
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/symptoms/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get a symptom log
      tags: [Symptoms]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Symptom log details
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update a symptom log
      tags: [Symptoms]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SymptomInput"
      responses:
        "200":
          description: Symptom log updated
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a symptom log
      tags: [Symptoms]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Symptom log deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ocd-session
  schemas:
...
    Symptom:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        anxietyLevel:
          type: integer
          minimum: 0
          maximum: 10
        compulsionCount:
          type: integer
        compulsionDuration:
          type: integer
        intrusiveThoughtFrequency:
          type: string
          enum: [none, rare, occasional, frequent, constant]
        resistanceLevel:
          type: integer
          minimum: 0
          maximum: 10
        mood:
          type: string
          enum: [great, good, neutral, bad, terrible]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
    SymptomInput:
      type: object
      required: [date, anxietyLevel]
      properties:
        date:
          type: string
        anxietyLevel:
          type: integer
          minimum: 0
          maximum: 10
        compulsionCount:
          type: integer
          default: 0
        compulsionDuration:
          type: integer
          default: 0
        intrusiveThoughtFrequency:
          type: string
          enum: [none, rare, occasional, frequent, constant]
          default: none
        resistanceLevel:
          type: integer
          minimum: 0
          maximum: 10
          default: 5
        mood:
          type: string
          enum: [great, good, neutral, bad, terrible]
          default: neutral
        notes:
          type: string
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Not found
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string